Moved identity keys data structure and functions to separate file
